refactor: Refactorizacion de funciones extras
Se divido en otro endpoint las funcionalidades extras que usa el frontend
