+++ main.py
@@ -1,8 +1,8 @@
 from fastapi import FastAPI
-from fastapi.staticfiles import StaticFiles
 from db.db_config import create_tables
-from routers import render,reportes,usuarios
+from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
+from routers import render,reportes,usuarios,functions
 
 app = FastAPI(title="MVP - Automatizacion ETCAR", 
             description="Proeycto de automatizacion",
@@ -17,3 +17,4 @@ app.mount("/static", StaticFiles(directory="static"), name="static")
 app.include_router(render.router)
 app.include_router(reportes.router)
 app.include_router(usuarios.router)
+app.include_router(functions.router)
