+++ templates/reporte.html
@@ -273,8 +273,8 @@
                                 <option value="" disabled selected>Seleccione el tipo...</option>
                                 <option value="Preventivo">Preventivo</option>
                                 <option value="Correctivo">Correctivo</option>
-                                <option value="Instalación">Instalación</option>
-                                <option value="Revisión">Revisión</option>
+                                <option value="Instalacion">Instalación</option>
+                                <option value="Revision">Revisión</option>
                             </select>
                         </div>
                     </div>
